+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [3.7.0](https://github.com/google-github-actions/release-please-action/compare/v3.6.1...v3.7.0) (2022-12-08)
+
+
+### Features
+
+* add `pull-request-header` input ([#615](https://github.com/google-github-actions/release-please-action/issues/615)) ([9963ea2](https://github.com/google-github-actions/release-please-action/commit/9963ea2aa819d5b257e4e6d1d66b82b25b7bafea))
+
+
+### Bug Fixes
+
+* update `release-please` to `14.17.5` ([#671](https://github.com/google-github-actions/release-please-action/issues/671)) ([4da3bc1](https://github.com/google-github-actions/release-please-action/commit/4da3bc15340c3f885c6200feee80e1f458213e15))
+
+## [3.6.1](https://github.com/google-github-actions/release-please-action/compare/v3.6.0...v3.6.1) (2022-11-23)
+
+
+### Bug Fixes
+
+* bump release-please from 14.16.0 to 14.17.2 ([#664](https://github.com/google-github-actions/release-please-action/issues/664)) ([c8f107f](https://github.com/google-github-actions/release-please-action/commit/c8f107f4cfb84600d0985a304bd17c16d7e4fac2))
+
+## [3.6.0](https://github.com/google-github-actions/release-please-action/compare/v3.5.1...v3.6.0) (2022-11-09)
+
+
+### Features
+
+* support proxy configuration ([#621](https://github.com/google-github-actions/release-please-action/issues/621)) ([714f07f](https://github.com/google-github-actions/release-please-action/commit/714f07f3ae692b5ce291aedf60f12280f4702eb2))
+
+
+### Bug Fixes
+
+* bump release-please from 14.14.0 to 14.16.0 ([#658](https://github.com/google-github-actions/release-please-action/issues/658)) ([a76b467](https://github.com/google-github-actions/release-please-action/commit/a76b4670614a5b05053bfd33aac25b3ffcb52287))
+
 ## [3.5.1](https://github.com/google-github-actions/release-please-action/compare/v3.5.0...v3.5.1) (2022-10-26)